+++ src/main/java/com/example/lawsearch/repository/LawRepository.java
@@ -7,8 +7,10 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface LawRepository extends JpaRepository<Law, Long> {
-
+public interface LawRepository extends JpaRepository<Law, Integer > {
+ // caut dupa titlu
     List<Law> findByTitleContainingIgnoreCase(String title);
+
+    //caut legi dupa o anumita versiune
     List<Law> findByVersion(String version);
 }
